+++ Source/dal/payments/delete-by-interval.js
@@ -1,14 +1,14 @@
 /**
- * Created by Ga5Xz2 on 31.12.2015..
+ * Created by I.Denisovs on 31.12.2015..
  */
 
 var db = require('../db');
 
-var sql = 'DELETE FROM payments WHERE time BETWEEN $start AND $end';
+var sql = 'DELETE FROM payments WHERE time BETWEEN $start AND $end AND userId = $userId';
 
-function deleteByInterval(interval, callback)
+function deleteByInterval(interval, userId, callback)
 {
-    var params = { $start: interval.start, $end: interval.end };
+    var params = { $start: interval.start, $end: interval.end, $userId: userId };
 
     db.run(sql, params, done);
